Link spicy-config to real compiler instead of Homebrew shim
By default Homebrew compiles with `CMAKE_CXX_COMPILER` set to the
Homebrew shim compiler. This e.g., injects extra steps to make sure the
shim is only executed by in a Homebrew-controlled environment, most
likely for exclusive use during building via Homebrew.

Since we want to be able to execute the compiler pointed to by
`spicy-config --cxx` outside of Homebrew (e.g., to build standalone
executables with `spicy-build`) we need to build Spicy with a real
compiler as well.